backend/requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
psycopg2-binary==2.9.5
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.6.0
requests==2.28.1
gunicorn==20.1.0
Flask-Cors==3.0.10
pydantic==1.10.2
